[IMPROVE] Improve loan table feature

+++ src/com/aptech/LoanProcessingSystem/model/LoanAndFineHistoryModel.java
@@ -15,8 +15,8 @@ import com.aptech.LoanProcessingSystem.entities.MyLoanAndFineHistory;
 
 public class LoanAndFineHistoryModel {
 
-	public List<MyLoanAndFineHistory> getAllLoanAndFineHistorys() {
-		List<MyLoanAndFineHistory> loanAndFineHistorys = null;
+	public List<LoanAndFineHistory> getAllLoanAndFineHistorys() {
+		List<LoanAndFineHistory> loanAndFineHistorys = null;
 		try {
 			loanAndFineHistorys = new ArrayList<>();
 			PreparedStatement statement = ConnectDB.connection()
@@ -26,11 +26,11 @@ public class LoanAndFineHistoryModel {
 							+ "left join paymentmethod d on d.id = a.PaymentMethodId");
 			ResultSet resultset = statement.executeQuery();
 			while (resultset.next()) {
-				MyLoanAndFineHistory loanAndFineHistory = new MyLoanAndFineHistory();
+				LoanAndFineHistory loanAndFineHistory = new LoanAndFineHistory();
 				loanAndFineHistory.setId(resultset.getInt("Id"));
 				loanAndFineHistory.setCustomer(resultset.getString("CName"));
 				loanAndFineHistory.setAmount(resultset.getDouble("Amount"));
-				loanAndFineHistory.setPaymentMethod(resultset.getString("PMName"));
+				loanAndFineHistory.setPaymentMenthodName(resultset.getString("PMName"));
 				loanAndFineHistory.setPaymentAmount(resultset.getDouble("PaymentAmount"));
 				loanAndFineHistory.setAmountLeft(resultset.getDouble("AmountLeft"));
 				loanAndFineHistory.setDueDate(resultset.getDate("DueDate"));
@@ -102,15 +102,30 @@ public class LoanAndFineHistoryModel {
 		return result;
 	}
 
+//	public boolean delete(int id) {
+//		boolean result = true;
+//		try {
+//			PreparedStatement preparedStatement = ConnectDB.connection()
+//					.prepareStatement("update LoanAndFineHistory SET status = false where id = ?");
+//			preparedStatement.setInt(1, id);
+//			result = preparedStatement.executeUpdate() > 0;
+//		} catch (Exception e) {
+//			result = false;
+//		} finally {
+//			ConnectDB.disconnect();
+//		}
+//		return result;
+//	}
+
 	public boolean delete(int id) {
-		boolean result = true;
+		boolean result = false;
 		try {
 			PreparedStatement preparedStatement = ConnectDB.connection()
 					.prepareStatement("delete from LoanAndFineHistory where id = ?");
 			preparedStatement.setInt(1, id);
 			result = preparedStatement.executeUpdate() > 0;
 		} catch (Exception e) {
-			result = false;
+			e.printStackTrace();
 		} finally {
 			ConnectDB.disconnect();
 		}
@@ -118,14 +133,27 @@ public class LoanAndFineHistoryModel {
 	}
 
 	public boolean deleteWithLoanId(int loanId) {
-		boolean result = true;
+		boolean result = false;
 		try {
-			PreparedStatement preparedStatement = ConnectDB.connection()
-					.prepareStatement("delete from LoanAndFineHistory where loanId = ?");
-			preparedStatement.setInt(1, loanId);
-			result = preparedStatement.executeUpdate() > 0;
+
+			PreparedStatement psCheck = ConnectDB.connection()
+					.prepareStatement("select count(id) as Count FROM LoanAndFineHistory where loanId = ?");
+			psCheck.setInt(1, loanId);
+			ResultSet resultSet = psCheck.executeQuery();
+			if (resultSet.next()) {
+				int count = resultSet.getInt("Count");
+				if (count == 0) {
+					result = true;
+				} else {
+					PreparedStatement psDelete = ConnectDB.connection()
+							.prepareStatement("delete from LoanAndFineHistory where loanId = ?");
+					psDelete.setInt(1, loanId);
+					result = psDelete.executeUpdate() > 0;
+				}
+			}
+
 		} catch (Exception e) {
-			result = false;
+			e.printStackTrace();
 		} finally {
 			ConnectDB.disconnect();
 		}
@@ -158,6 +186,7 @@ public class LoanAndFineHistoryModel {
 				loanAndFineHistory = null;
 			}
 		} catch (Exception e) {
+			e.printStackTrace();
 			loanAndFineHistory = null;
 		} finally {
 			ConnectDB.disconnect();
@@ -187,8 +216,8 @@ public class LoanAndFineHistoryModel {
 		return result;
 	}
 
-	public List<MyLoanAndFineHistory> search(String keyword) {
-		List<MyLoanAndFineHistory> loanAndFineHistorys = null;
+	public List<LoanAndFineHistory> search(String keyword) {
+		List<LoanAndFineHistory> loanAndFineHistorys = null;
 		try {
 			loanAndFineHistorys = new ArrayList<>();
 			PreparedStatement statement = ConnectDB.connection()
@@ -199,11 +228,49 @@ public class LoanAndFineHistoryModel {
 			statement.setString(1, "%" + keyword + "%");
 			ResultSet resultset = statement.executeQuery();
 			while (resultset.next()) {
-				MyLoanAndFineHistory loanAndFineHistory = new MyLoanAndFineHistory();
+				LoanAndFineHistory loanAndFineHistory = new LoanAndFineHistory();
 				loanAndFineHistory.setId(resultset.getInt("Id"));
 				loanAndFineHistory.setCustomer(resultset.getString("CName"));
 				loanAndFineHistory.setAmount(resultset.getDouble("Amount"));
-				loanAndFineHistory.setPaymentMethod(resultset.getString("PMName"));
+				loanAndFineHistory.setPaymentMenthodName(resultset.getString("PMName"));
+				loanAndFineHistory.setPaymentAmount(resultset.getDouble("PaymentAmount"));
+				loanAndFineHistory.setAmountLeft(resultset.getDouble("AmountLeft"));
+				loanAndFineHistory.setDueDate(resultset.getDate("DueDate"));
+				loanAndFineHistory.setFineInterest(resultset.getFloat("FineInterest"));
+				loanAndFineHistory.setFineOverDays(resultset.getInt("FineOverDays"));
+				loanAndFineHistory.setFineAmount(resultset.getDouble("FineAmount"));
+				loanAndFineHistory.setPaymentDate(resultset.getDate("PaymentDate"));
+				loanAndFineHistory.setDescription(resultset.getString("Description"));
+				loanAndFineHistory.setStatus(resultset.getBoolean("Status"));
+				loanAndFineHistorys.add(loanAndFineHistory);
+			}
+		} catch (Exception e) {
+			e.printStackTrace();
+			loanAndFineHistorys = null;
+		} finally {
+			ConnectDB.disconnect();
+		}
+		return loanAndFineHistorys;
+	}
+
+	public List<LoanAndFineHistory> searchByLoanId(int loanId) {
+
+		List<LoanAndFineHistory> loanAndFineHistorys = null;
+		try {
+			loanAndFineHistorys = new ArrayList<>();
+			PreparedStatement statement = ConnectDB.connection()
+					.prepareStatement("Select a.*, c.Name as 'CName', d.Name as 'PMName' from LoanAndFineHistory a\r\n"
+							+ "left JOIN loan b on a.LoanId = b.Id\r\n"
+							+ "left join customer c on c.id = b.CustomerId\r\n"
+							+ "left join paymentmethod d on d.id = a.PaymentMethodId\r\n " + "where a.LoanId = ?");
+			statement.setInt(1, loanId);
+			ResultSet resultset = statement.executeQuery();
+			while (resultset.next()) {
+				LoanAndFineHistory loanAndFineHistory = new LoanAndFineHistory();
+				loanAndFineHistory.setId(resultset.getInt("Id"));
+				loanAndFineHistory.setCustomer(resultset.getString("CName"));
+				loanAndFineHistory.setAmount(resultset.getDouble("Amount"));
+				loanAndFineHistory.setPaymentMenthodName(resultset.getString("PMName"));
 				loanAndFineHistory.setPaymentAmount(resultset.getDouble("PaymentAmount"));
 				loanAndFineHistory.setAmountLeft(resultset.getDouble("AmountLeft"));
 				loanAndFineHistory.setDueDate(resultset.getDate("DueDate"));
